# diploma_work/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Structure API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Registration",
                "description": "Register a new user and get a JWT.",
                "operationId": "ca7c0cb6772de9bffbb8c896980cebc8",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "user"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user230@crocos.kz"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "your_token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Could not create user"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Login",
                "description": "Log in a user and get a JWT.",
                "operationId": "c373bc4cffa4a56fc64e3979b8ee7b96",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user230@crocos.kz"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "your_token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "These credentials do not match our records."
                    },
                    "500": {
                        "description": "Could not create token"
                    }
                }
            }
        },
        "/api/forgot": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Forgot password",
                "description": "Handle forgot password request.",
                "operationId": "ea1a70c23193e0d3eb9113fcba2e1575",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user230@crocos.kz"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "We have emailed your password reset link."
                    },
                    "500": {
                        "description": "Could not send mail or could not create reset code"
                    }
                }
            }
        },
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Reset password",
                "description": "Reset user password.",
                "operationId": "f3732be9b471fb1d0f6149c860ae5356",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "reset_code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user230@crocos.kz"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "reset_code": {
                                        "type": "string",
                                        "example": "f28adw"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Your password has been reset."
                    },
                    "500": {
                        "description": "Could not change password"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Logout",
                "description": "Log out the authenticated user.",
                "operationId": "5ff2dc8b48bff52908d3821a0ecc9b3e",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "example": "Bearer your_token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    },
                    "404": {
                        "description": "Error message"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RunningSessionResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "distance_km": {
                        "type": "number",
                        "format": "float",
                        "example": 5.5
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-24T14:00:00Z"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-24T14:30:00Z"
                    },
                    "total_time": {
                        "type": "integer",
                        "example": 1800
                    },
                    "average_speed": {
                        "type": "number",
                        "format": "float",
                        "example": 10.5
                    },
                    "max_speed": {
                        "type": "number",
                        "format": "float",
                        "example": 12.5
                    },
                    "calories_burned": {
                        "type": "integer",
                        "example": 300
                    },
                    "points": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "test"
                    },
                    "email": {
                        "type": "string",
                        "example": "example@example.com"
                    },
                    "profile_picture": {
                        "type": "string",
                        "example": "profile_pictures/adagasf21412tgweg"
                    },
                    "weight_kg": {
                        "type": "number",
                        "format": "float",
                        "example": 62.5
                    },
                    "height_cm": {
                        "type": "number",
                        "format": "float",
                        "example": 181
                    },
                    "points": {
                        "type": "integer",
                        "example": 50
                    },
                    "running_sessions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RunningSessionResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-24T14:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-24T14:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}